requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.6.1